backend/requirements.txt
Flask>=2.0
Flask-SocketIO>=5.3.2
python-socketio>=5.7.2
eventlet>=0.33.0
Flask-SQLAlchemy>=3.0
python-dotenv>=0.20.0
Flask-Cors>=3.0.10
mysql-connector-python